MC-15776: Merge release branch into 2.3-develop
- fix merge conflicts

+++ app/code/Magento/QuoteGraphQl/Model/Resolver/UpdateCartItems.php
@@ -110,10 +110,45 @@ class UpdateCartItems implements ResolverInterface
             }
             $quantity = (float)$item['quantity'];
 
+            $cartItem = $cart->getItemById($itemId);
+            if ($cartItem === false) {
+                throw new GraphQlNoSuchEntityException(
+                    __('Could not find cart item with id: %1.', $item['cart_item_id'])
+                );
+            }
+
             if ($quantity <= 0.0) {
                 $this->cartItemRepository->deleteById((int)$cart->getId(), $itemId);
             } else {
-                $this->updateCartItem->execute($cart, $itemId, $quantity, $customizableOptions);
+                $cartItem->setQty($quantity);
+                $this->validateCartItem($cartItem);
+                $this->cartItemRepository->save($cartItem);
+            }
+        }
+    }
+
+    /**
+     * Validate cart item
+     *
+     * @param Item $cartItem
+     * @return void
+     * @throws GraphQlInputException
+     */
+    private function validateCartItem(Item $cartItem): void
+    {
+        if ($cartItem->getHasError()) {
+            $errors = [];
+            foreach ($cartItem->getMessage(false) as $message) {
+                $errors[] = $message;
+            }
+
+            if (!empty($errors)) {
+                throw new GraphQlInputException(
+                    __(
+                        'Could not update the product with SKU %sku: %message',
+                        ['sku' => $cartItem->getSku(), 'message' => __(implode("\n", $errors))]
+                    )
+                );
             }
         }
     }
